core/standards/css3/image-fit/js/inherited/004.html

<!doctype html>
<title>'inherit' keyword for object-position</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div style='-o-object-position:0% 0%'><div style='-o-object-position:inherit'></div></div>
<script>
test(function() {
  var e = document.body.firstChild.firstChild;
  assert_equals(getComputedStyle(e, '').OObjectFit, 'auto');
  assert_equals(getComputedStyle(e, '').OObjectPosition, '0% 0%');
});
</script>
